Rename project to identity-local-auth-risk
