--- components/org.wso2.carbon.identity.adaptive.authentication.riskscore/src/main/java/org/wso2/carbon/identity/adaptive/authentication/riskscore/util/RiskScoreDTO.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.wso2.carbon.identity.adaptive.authentication.riskscore.util;
-
-/**
- * Contains the Risk score object returned from the IS Analytics
- */
-public class RiskScoreDTO {
-    private int score;
-
-    public RiskScoreDTO() {
-    }
-
-    public RiskScoreDTO(int score) {
-        this.score = score;
-    }
-
-    public int getScore() {
-        return score;
-    }
-
-    public void setScore(int score) {
-        this.score = score;
-    }
-}
