+++ components/org.wso2.carbon.identity.authenticator.risk/src/main/java/org/wso2/carbon/identity/authenticator/risk/internal/JavascriptFunctionsSupplierComponent.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.wso2.carbon.identity.authenticator.risk.internal;
+
+import org.osgi.service.component.ComponentContext;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
+import org.osgi.service.component.annotations.ReferencePolicy;
+import org.wso2.carbon.identity.authenticator.risk.GetRiskScoreFunction;
+import org.wso2.carbon.identity.authenticator.risk.GetRiskScoreFunctionImpl;
+import org.wso2.carbon.identity.application.authentication.framework.JsFunctionRegistry;
+
+/**
+ * TODO: Class level comments
+ */
+@Component(
+        name = "identity.application.authentication.framework.extension.js.riskscore.functions.component",
+        immediate = true)
+public class JavascriptFunctionsSupplierComponent {
+
+    private JsFunctionRegistry jsFunctionRegistry;
+    private GetRiskScoreFunctionImpl getRiskScoreFunction;
+
+    @Activate
+    protected void activate(ComponentContext ctxt) {
+        getRiskScoreFunction = new GetRiskScoreFunctionImpl();
+        jsFunctionRegistry.register(JsFunctionRegistry.Subsystem.SEQUENCE_HANDLER, "getRiskScore",
+                (GetRiskScoreFunction) getRiskScoreFunction::getRiskScore);
+    }
+
+    @Deactivate
+    protected void deactivate(ComponentContext ctxt) {
+        if (jsFunctionRegistry != null) {
+            jsFunctionRegistry.deRegister(JsFunctionRegistry.Subsystem.SEQUENCE_HANDLER, "getRiskScore");
+        }
+    }
+
+    @Reference(
+            service = JsFunctionRegistry.class,
+            cardinality = ReferenceCardinality.MANDATORY,
+            policy = ReferencePolicy.DYNAMIC,
+            unbind = "unsetJsFunctionRegistry"
+    )
+    public void setJsFunctionRegistry(JsFunctionRegistry jsFunctionRegistry) {
+        this.jsFunctionRegistry = jsFunctionRegistry;
+    }
+
+    public void unsetJsFunctionRegistry(JsFunctionRegistry jsFunctionRegistry) {
+        this.jsFunctionRegistry = null;
+    }
+}
