+++ pom.xml
@@ -20,22 +20,26 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.wso2.carbon.identity.adaptive.authentication</groupId>
-    <artifactId>adaptive-authentication</artifactId>
-    <packaging>pom</packaging>
-    <version>1.0-SNAPSHOT</version>
-    <name>WSO2 Carbon - Adaptive Authenticator</name>
-    <url>http://wso2.org</url>
 
     <parent>
         <groupId>org.wso2</groupId>
         <artifactId>wso2</artifactId>
         <version>1</version>
     </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>org.wso2.carbon.identity.local.auth.risk</groupId>
+    <artifactId>identity-local-auth-risk</artifactId>
+    <packaging>pom</packaging>
+    <version>1.0-SNAPSHOT</version>
+    <name>WSO2 Carbon - Risk Calculator for Authentication Request </name>
+    <url>http://wso2.org</url>
+
+
     <modules>
-        <module>components/org.wso2.carbon.identity.adaptive.authentication.riskscore</module>
+        <module>components/org.wso2.carbon.identity.authenticator.risk</module>
+        <module>features/org.wso2.carbon.identity.authenticator.risk.analytics.feature</module>
+        <module>components/org.wso2.carbon.identity.authenticator.risk.analytics.endpoint</module>
     </modules>
 
     <repositories>
